refactor: simplified logic of <CreateRequest> component
